Rename model package to entity
