+++ src/main/java/com/bytebreeze/quickdrop/entity/User.java
@@ -1,4 +1,4 @@
-package com.bytebreeze.quickdrop.model;
+package com.bytebreeze.quickdrop.entity;
 
 import com.bytebreeze.quickdrop.enums.EmailVerified;
 import com.bytebreeze.quickdrop.enums.Role;
