+++ src/main/java/com/bytebreeze/quickdrop/repository/ParcelRepository.java
@@ -1,8 +1,8 @@
 package com.bytebreeze.quickdrop.repository;
 
 import com.bytebreeze.quickdrop.enums.ParcelStatus;
-import com.bytebreeze.quickdrop.model.Parcel;
-import com.bytebreeze.quickdrop.model.Rider;
+import com.bytebreeze.quickdrop.entity.Parcel;
+import com.bytebreeze.quickdrop.entity.Rider;
 import java.util.List;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
