+++ src/main/java/com/bytebreeze/quickdrop/repository/ProductCategoryRepository.java
@@ -1,6 +1,6 @@
 package com.bytebreeze.quickdrop.repository;
 
-import com.bytebreeze.quickdrop.model.ProductCategory;
+import com.bytebreeze.quickdrop.entity.ProductCategory;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
