+++ src/main/java/com/bytebreeze/quickdrop/service/ParcelService.java
@@ -6,7 +6,7 @@ import com.bytebreeze.quickdrop.dto.request.CalculateShippingCostRequestDto;
 import com.bytebreeze.quickdrop.dto.request.ParcelBookingRequestDTO;
 import com.bytebreeze.quickdrop.enums.ParcelStatus;
 import com.bytebreeze.quickdrop.enums.PaymentStatus;
-import com.bytebreeze.quickdrop.model.*;
+import com.bytebreeze.quickdrop.entity.*;
 import com.bytebreeze.quickdrop.repository.*;
 import com.bytebreeze.quickdrop.util.AuthUtil;
 import jakarta.transaction.Transactional;
