+++ src/main/java/com/bytebreeze/quickdrop/service/RiderService.java
@@ -14,8 +14,8 @@ import com.bytebreeze.quickdrop.exception.custom.ParcelAlreadyAssignedException;
 import com.bytebreeze.quickdrop.exception.custom.ParcelNotFoundException;
 import com.bytebreeze.quickdrop.exception.custom.UserNotFoundException;
 import com.bytebreeze.quickdrop.mapper.RegisterRiderMapper;
-import com.bytebreeze.quickdrop.model.Parcel;
-import com.bytebreeze.quickdrop.model.Rider;
+import com.bytebreeze.quickdrop.entity.Parcel;
+import com.bytebreeze.quickdrop.entity.Rider;
 import com.bytebreeze.quickdrop.repository.ParcelRepository;
 import com.bytebreeze.quickdrop.repository.RiderRepository;
 import com.bytebreeze.quickdrop.repository.UserRepository;
