+++ src/test/java/com/bytebreeze/quickdrop/controller/RiderControllerTest.java
@@ -12,8 +12,8 @@ import com.bytebreeze.quickdrop.enums.ParcelStatus;
 import com.bytebreeze.quickdrop.enums.VerificationStatus;
 import com.bytebreeze.quickdrop.exception.custom.AlreadyExistsException;
 import com.bytebreeze.quickdrop.exception.custom.ParcelNotFoundException;
-import com.bytebreeze.quickdrop.model.Parcel;
-import com.bytebreeze.quickdrop.model.Rider;
+import com.bytebreeze.quickdrop.entity.Parcel;
+import com.bytebreeze.quickdrop.entity.Rider;
 import com.bytebreeze.quickdrop.service.ParcelService;
 import com.bytebreeze.quickdrop.service.RiderService;
 import java.math.BigDecimal;
