+++ src/test/java/com/bytebreeze/quickdrop/service/ParcelServiceTest.java
@@ -7,11 +7,11 @@ import static org.mockito.Mockito.*;
 import com.bytebreeze.quickdrop.dto.request.CalculateShippingCostRequestDto;
 import com.bytebreeze.quickdrop.dto.request.ParcelBookingRequestDTO;
 import com.bytebreeze.quickdrop.enums.ParcelStatus;
-import com.bytebreeze.quickdrop.model.Parcel;
-import com.bytebreeze.quickdrop.model.Payment;
-import com.bytebreeze.quickdrop.model.ProductCategory;
-import com.bytebreeze.quickdrop.model.Rider;
-import com.bytebreeze.quickdrop.model.User;
+import com.bytebreeze.quickdrop.entity.Parcel;
+import com.bytebreeze.quickdrop.entity.Payment;
+import com.bytebreeze.quickdrop.entity.ProductCategory;
+import com.bytebreeze.quickdrop.entity.Rider;
+import com.bytebreeze.quickdrop.entity.User;
 import com.bytebreeze.quickdrop.repository.ParcelRepository;
 import com.bytebreeze.quickdrop.repository.PaymentRepository;
 import com.bytebreeze.quickdrop.repository.ProductCategoryRepository;
